new StationInitializer in own module and more tests
